The new API refactoring
