+++ src/test/java/io/appium/java_client/android/SettingTest.java
@@ -3,20 +3,17 @@ package io.appium.java_client.android;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import io.appium.java_client.android.settings.Setting;
 import org.junit.Test;
 
 public class SettingTest extends BaseAndroidTest {
 
     @Test public void ignoreUnimportantViewsTest() {
-        driver.setSetting(Setting.IGNORE_UNIMPORTANT_VIEWS,
-                Setting.IgnoreUnimportantViews.YES.value());
+        driver.ignoreUnimportantViews(true);
         boolean ignoreViews =
                 driver.getSettings().get(Setting.IGNORE_UNIMPORTANT_VIEWS.toString())
                         .getAsBoolean();
         assertTrue(ignoreViews);
-        driver.setSetting(Setting.IGNORE_UNIMPORTANT_VIEWS,
-                Setting.IgnoreUnimportantViews.NO.value());
+        driver.ignoreUnimportantViews(false);
         ignoreViews = driver.getSettings().get(Setting.IGNORE_UNIMPORTANT_VIEWS.toString())
                 .getAsBoolean();
         assertFalse(ignoreViews);
